// client/content-components/routes/routes.ts
export enum ERoute {
  SignIn = "sign-in",
  SecurePassword = "secure-password",
  Campaigns = "campaigns",
  Wallet = "wallet",
  Deposit = "deposit",
  Withdraw = "withdraw",
  Activity = "activity",
  More = "more",
  SecurityAndPrivacy = "security-and-privacy",
  PrivateKey = "private-key",
  ChangePassword = "change-password"
}